Add additional visual-transformer scripts
